requirements.txt
metadict
pyroute2
python-dateutil
requests
termcolor
tmsyscall
tqdm
typer